Seperated story and restored original
